break up contexts
