+++ src/tabViews/window/WindowView.tsx
@@ -8,7 +8,7 @@ import {
   useTabGroups,
   useTabs,
   useWindow,
-} from "src/contexts/selectors";
+} from "src/contexts";
 import { filterSortTabs } from "src/utils/filterTabs";
 import TabGroupView from "./TabGroupView";
 import TabView from "./TabView";
